gx_material: Properly emulate TEV overflow stage clamping...
Oops. Turns out that only the lower 8 bits of inputs A, B & C
are used. Input D is a full 10-bit register, so it doesn't get overflow.

This fixes the water at the beginning of the Water Temple in
Twilight Princess... and also drastically changes the look of
the water effect in Super Mario Sunshine :(

I'll have to update the blog post about it with the new details.