gx: Introduce gx_render helpers
While it's already pretty easy to create a renderer
using gx_displaylist, gx_material and the folk,
the UBO update stuff can be a bit of a pain.

This adds some helpers to help with the more
annoying and tedious parts of creating the
rendering half of things.

Ultimately, this is more about uploading to GL
more than rendering, but shhhh...