render: Rewrite state management
When I was calling RenderState.flatten, I didn't realize it was
modifying the passed in render state flags... which means that
various bits of things can leak through depending on when
the RenderState flags are first applied.

Fixes the bizarre SM64DS rendering bugs that happen if you
switch from a GX game to it.