Remove zelview
This is almost temporary -- it's one of the last remaining
non-Gfx ports, it's unmaintained, and honestly OoT3D's
artwork is better in practically every way.

N.E.C.'s fixes to bring in texenv broke a lot of the scenes
and my attempts to fix were fruitless.

If N.E.C or someone else wants to bring it back and port it
to Gfx, be my guest, but I have very little desire to do so
myself.