Adding New Camera Speed Slider.
Adding a Camera Speed Slider to the viewer settings. I also hooked up
the slider to also update on the mouse wheel event, so that the camera
speed can still be set like it was previously, or by using this new
slider. Because of the addition of this slider, the max camera speed is
1000 normally, or 5000 if you are holding shift.