gfx: Add new UBO allocator strategy
... and start the porting with Katamari Damacy, as it's a relatively
recent renderer without too many dependenices.

The original plan for Gfx was to be similar to a "modern" system
where everything was explicitly allocated and bound up-front.
As the site developed, though, kludges like late-sampler-binding
meant that we needed to recreate GfxBindings. Render hacks meant
we needed dynamic GfxRenderPipeline building as well. And as the
time went on, more and more got pushed from the up-front
GfxRenderInstBuilder towards the runtime GfxRenderInsts themselves...

As we continue, we might want to add features like particle systems
or dynamic object spawns where we might not be able to create a
GfxRenderInst up-front at load, anyway... at least not without hacks.

The last remaining stronghold of preventing this been the uniform
buffer side of GfxBindings. It's uglier than it initially was because of
Galaxy's choice to prefer to spawn models after load time, but
GfxRenderInstBuilder was still a requirement in the scheme.

Start to break that down now. Instead of pre-allocating all UBOs into
separate buffers, we now use one giant UBO which we can partition
off, and can partition off once-per-frame.

The eventual idea is to remove GfxRenderInstBuilder in favor of an
object pool, but this will require a bit more refactoring so that we can
set this up easily across all systems.

But this is the broad plan, and the broad strokes of the refactor.

Namaste.