j3d: Clean up envelope effected stuff
I misread the code before and did it by hand,
but now I understand what calcPostTexMtx is
doing so I can move it to generic code.

... except this still isn't right, because reading
more tells me that J3D hasn't had code to
actually parse the posttexmtx's in a looong
time. But that's for another day :P