Camera: Reverse depth & use infinite far plane
In theory, this makes no sense, because OpenGL doesn't benefit at all
from reversed depth. But since WebGL is commonly implemented on
top of ANGLE, which is some emulation code around D3D11, suddenly
this makes sense, since some extra vertex shader code is used to handle
the differing NDC space:

https://github.com/google/angle/blob/1dbbc7b3e23296acc3892e317b1fe668bcb49585/src/libANGLE/renderer/d3d/DynamicHLSL.cpp#L576-L578

So despite GL technically having a -1...1 depth range, it actually gets
compressed down into 0...1 and thus this has a pretty decent impact
on most scenes.

Additionally, take the recommendation of Upchurch and Desbrun, and
use an infinite far plane. It means we don't have to manually set it
anymore, and it's not that bad for precision anyway...