Banjo Kazooie: Implement second particle system.
Add the second particle system, which has many more knobs to set in
each emitter. This handles most particles not connected to the player.
There doesn't seem to be a hard particle maximum on these, so the pool
should be larger.

Switch particle timer from frames to time, which seems more common.

Also fix for zero fps in my new animation controller method.