Pokemon Snap: Animate textures.
Implement almost all of the material animation logic. Lighting is still
missing, as is full support of scaling since I haven't gone through and
found examples that use it. Textures are loaded in advance rather than on
the fly partially as a sanity check for the format and loading logic.

Recognize (some?) billboard bits. I couldn't tell from the logic if it was
supposed to affect children, but it doesn't seem to matter.

Add basic opacity selection/sorting logic. More to come, unclear how
much the game does.

Improve handling of multi-DL models, which append to separate display
lists for opaque and translucent. In principle they could carry over
mode settings between items, but we reset between each for now.