SFA/DP ancient maps labeled
most of the ancient maps are DP versions of SFA maps and are labeled as such. the cloudrunner fortress levels are closer to retail SFA, so are labeled as early SFA. question marks are on ones i'm unsure of; there's nothing in SFA or existing DP footage that resembles the pit or mine rooms, so those are marked as "unknown".