Pokemon Snap: Run object state machines.
Aside from a couple of fixes to initial states which used switches,
everything is run as is. No motion logic is handled yet, nor is signal
passing. All animations are treated as loops, which looks wrong in some
cases - it's possible we could guess from the state end condition.

Put basic logic for interactivity in a new global object. This could
be exposed in a more direct way later.