Pokemon Snap: Finish volcano interactions.
Volcano is done for the foreseeable future.

Charmeleon's path turned out to be a slightly modified version of the
common path function. Volcano has a few local versions of global functions
for whatever reason.

The magmar/charmander interaction relies on collision, so I'm glad I had
that already. It's still not quite right, as in game it can be cancelled
by projectiles differently.

Also fix two spawns where the code incorrectly puts the ID of the lava
splash. This feels like more evidence that volcano was either done first
or by someone who wasn't aware of conventions in the code.