Pokemon Snap: Improve river and Zapdos.
Finish Psyduck, Poliwag, and Slowpoke, and also fix the logic for what
jumps out of the water. The first Psyduck isn't actually supposed to
loop, but this seemed better given the nature of the viewer.

Zapdos can now hatch. I had forgotten about the egg shattering effect,
which will take a bit more work to implement.

Fix an old bug, where the alpha of lit vertices was set to 1. In most
cases this was probably fine, but Electrode's explosion actually relied
on the vertex alpha being less than 1, and was invisible. Related to
this, make the alpha visualization use the final shaded alpha instead of
the vertex alpha, so now everything should be visible with that mode.

Also fix two smaller issues with animations: always reset animator
values to the original node values (I had misread a branch condition
before) and update the collision center when the root is animated. The
latter was responsible for lots of buggy-looking projectile misses.