Pokemon Snap: Implement egg shatter effect.
Two of the legendary bird eggs shatter by interpolating position between
their main vertex data and an additional buffer. Normals are also
interpolated, though this doesn't make physical sense.

To minimize the impact on existing rendering code, I just added a new
vertex buffer to the shader input and prepended the interpolation to the
vertex shader. Only the attributes that change are saved in the new buffer.