Pokemon Snap: Add last two evolutions.
Now every pokemon in the game is present.

Hack in support for Grimer's separate material animation speed.

Make maintaining current phase the default for the adjustable animation
controller, rather than resetting to zero. It's just easier to do inside
the class.

Fix an error in I8 texture padding and replace all computations with a
function. Magneton's eye seems to be the first oddly-sized I8 texture.