Banjo Tooie: Start adding objects from setup files.
Map setup files are similar to B-K, but are encrypted using the CIC.
Most actors have their own DLL, so need to use another DLL to point us to
the right place for each ID. Static objects and collectibles have their own
special ID lookup tables. Similar to B-K, the list of objects includes
IDs that mean several different things, so it's hard to tell how much more
logic will be needed.

Now that more models are loaded, it's clear there's work to be done on
the geo layout commands. Flipbooks are easier to detect now, but are
just skipped for now.