GfxRenderer: Introduce late sampler binding system
Sampler binding + render target + render pass management
has always been tricky. The conflicting issue here is that we
want the ability to create all GfxRenderInst's up front for
single-upload uniform buffer reasons, but we also want the
flexibility to allocate RTs at a later stage, or bind different
textures to the same "abstract texture" based on per-pass
semantics.

Ultimately, I came to the decision that it would make sense
for the late binding system to be part of GfxRenderer, rather
than any app code, since Gfx is the one enforcing this uniform
buffer constraint, and so it should make concessions to help
that aim.

So GfxRenderer now has an abstract "late sampler binding"
system, which is intended to be used with RTs. Additionally,

The API isn't perfect -- in an ideal world, no part of GfxPlatform
would have to change and this could be done at the GfxRenderer
layer. But for now, GfxSamplerBinding takes on an "additional"
field for practicality reasons. The API can be cleaned up later.

Switch over Galaxy's renderer to use the new late binding system.