ui: Don't change cursor / pointer-events if unnecessary
With large UI scenes (like the textures panel in Pokemon
HGSS), Chrome has to propagate restyles down to
every element in the DOM tree. See:

https://bugs.chromium.org/p/chromium/issues/detail?id=664066

Thankfully, in the majority case, pointer lock will
already hide the cursor and route events to the
locked element, so we can luckily avoid this.

Another day in "web scale".