Fix a few UI quibbles.
Disallow blank or whitespace keyframe names as this leaves the keyframes unclickable. Turn off autocomplete for keyframe names because it's distracting. Fetch interpolation type when selecting a keyframe even if the keyframe currently has a duration of 0 seconds (because changing that will display the buttons, with the last interpolation type selected).