gx_displaylist: Rename LoadedVertexPacket
to LoadedVertexDraw, since I want to remove the
inaccurate "Packet" terminology going forward.
Though "PacketParams" will be a larger one to
change, so that stays for now. Maybe in the future
we can even auto-optimize it to a single matrix!

Add some documentation around its intentions
and conventions, as well.