Katamari Damacy: Implement most remaining path motions.
There were a few more motion types involving paths which are used in the
game. Almost all of the unimplemented logic only becomes active when an
object is smaller than the player and trying to escape, so it's not
entirely clear how we should approach them.

The conditions for switching to alternate motions are now more accurate,
modulo some artistic license in how they should be interpreted. I only
consider the camera to be "in" the zone if it is sufficiently close to
the ground, and picked an arbitrary small cutoff size for objects to be
scared of the camera. We might want to tie that to the current area.

Also fix an embarrassing bug with loading path data and fix the edge
case of turning logic when turning exactly 180 degrees.

I shuffled some motion names around, hopefully makes more sense now.