Katamari Damacy: Add special tutorial logic.
Load the extra model and spawn data for the tutorial, and implement its
ad hoc animations. This required implementing the overall alpha multiplier,
which is also used to fade out distant objects.

It's definitely messy due to plumbing the tutorial bool through
everything, but this is likely the last major change to scene code.