Katamari Damacy: Clarify texture buffer format.
Texture buffers have inconsistent length because they are actually
chains of DMAtags. So far this only mattered for level textures, but
some menu models rely on this as well.

Also check the vertex color flag (though don't do anything with the
colors for now) and skip nil CLUTs, which fixes the credits crash.

Fix a crash I introduced a while ago due to level 27 not having
collision for some objects.