GfxRenderer: Move the shadow buffer allocation to allocate time
In theory, the separation between the chunk allocation and the
shadow buffer allocation could give us a bit of extra leeway if
the user allocated multiple chunks in a row. In practice, that's
unlikely to have happened, and in fact we have a bit of a footgun
for an API: If someone allocates a buffer but doesn't grow the
buffer for shadow space (say they allocate but never use it),
we can crash.

Just move the shadow buffer allocation to be inline with the
chunk allocation. It will get amortized over the rest of the
frame anyway.