viewer: Increase default FoV from 45 to 60 degrees
Low FoV values such as 45 are common in console games where the TV
screen is usually quite far, but not so in Desktop/mobile scenarioes
where noclip.website is most commonly used for.

Values of 80-90 are much more common in these cases, I picked 60 since
it's around the middle (also it divided evenly).

Signed-off-by: Nick Renieris <velocityra@gmail.com>