Final Fantasy X: Extract official location names.
A couple of the official names probably aren't ever seen and so don't make
much sense (in particular, one of the "Airship - Deck" I put under
"Fighting Sin" is just a model of Sin).

I've added disambiguation to a few of the cases with identical names,
but others will have to wait until textures. The names also make clear
that the commented-out, assert-breaking scenes, as well as scenes which
appear to be a single model, are fixed camera rooms with limited geometry.