FFX: Start interpreting event bytecode.
Parse and run a small subset of bytecode associated with "events", which
seems to encompass both normal levels and cutscenes/actor interactions.
In particular, actor spawn placement happens dynamically through the
scripts.

The current implementation is enough for doors and other mostly-static
objects, as well as a few rotating objects which stop under certain
conditions. While some opcodes relating to accessing level data remain
unimplemented, the bulk of the work that's left is supporting more of
the (several hundred) functions that scripts can call.