GX: Remove NBT from GX.Attr
It's always been a front-end fake in GX, and though
I tried to make it work, I just couldn't get around
it being in the wrong order in the display list,
and other weird differences that make it implusible to support.

So just drop the act, and remove the NBT fake,
and fix code to adjust.