Scene Graph: Try out an alternate API for pass construction
Not sure how I feel about it yet -- the callback makes it less clear
that stuff is happening immediately.

I like that pass objects are "pure" now, but without the setup func,
it isn't obvious that the pass objects are temporary and shouldn't
really be kept around... thankfully this is all ergonomics now and
we can easily change this. The fundamentals are here.