gfx: Remove GfxHostAccessPass
This is an absolutely massive change, woo!

This is now replaced with a device method
which "executes on the GPU timeline", i.e.
an implicitly queued copyBufferToBuffer,
which matches the OpenGL sematnics.

I might change things in the future if I choose
to go with an editlist-style API. It will depend
how WebGPU evolves going forward.