Source Engine: Properly setup SRGB state
The VTF flag itself doesn't do anything, instead we have to pay
attention to the SRGB sampler state set by the material. In D3D,
that can be adjusted with a sampler state, which is easy. In desktop GL,
we could use a glTextureView, but in GL ES, we don't have such
privileges, so we have to upload an entire new texture. Unfortunate,
but hey, what are you gonna do...

Fixes brickwall006, as seen in cp_sunshine.