Source: Refactor how materials are created
Move the init() back to the caller helps it set
up extra parameters without requiring a lot
of arguments to createMaterialInstance. Especially
since a few of these options, namely entityParams,
should be set *before* init() for certain operations
to work correctly.

This fixes texture toggles on certain surfaces not
working correctly, namely the signage posted along
the pipe in sp_a3_00. Note that there's some other
fun stuff that makes that map work correctly,
this just fixes the texturetoggle there...

This also lets us move the skinning mode to not
be a shader recompile.