J3D: Fix loading of SuperBMD models
At least some versions of SuperBMD produced models
without the EVP1 double-count bug. While this
means that they would crash in Super Mario Galaxy,
earlier versions of this engine would do not correct
for this behavior, and so they would fine in those games.

Tolerate this behavior by checking for the existence
of a SuperBMD-created model.