Source: Clean up BSPFile
Adjust terminology. For a long time, "surfaces" were parsed and
then merged, but it meant that it was tricky to determine what
was an "unmerged" surface and a "surface", and indexes were
confusing. Just adapt BSPFile's "face" terminology for BSP faces,
and use "Surface" strictly for rendering. So, e.g. surfleaflist now
becomes faceleaflist.

Should hopefully be cleaner to read now.