Super Mario Sunshine: Initial implementation for object lighting
The game first organizes light sources into four TLightWithDBSet-s:

  0: TPlayerLightWithDBSet
  1: TObjectLightWithDBSet
  2: TMapObjectLightWithDBSet
  3: TIndirectLightWithDBSet

Each of these search by name into the LightAry and AmbAry loaded
from the scene.bin. However, TObject…, TMapObject…, and TIndirect…
all search for the same name, so they are collapsed for this
implementation. Additionally, the index into the LightAry/AmbAry
is adjusted when an object is in shadow, but that is not yet
implemented here.

Also some scene names from stagename.bmg and scenarioname.bmg.