DebugTextDrawer: Fix cache issues between games
We were using DataShare for the DebugTextDrawer, while using the
GfxRenderInstManager's cache in the shared data... the
GfxRenderInstManager cache is not guaranteed to be a stable cache
between map changes.

Fix this by providing our own cache. This makes the DDraw/Font API
a bit uglier; I'll try to fix this later, but for now, this is what I'm doing :/