Added d_wood.ts. Basic bush rendering.
This is the start of one of the remaining missing d_s_play components, Wood (which is the TWW misnomer for bushes).

Instead of following the old Grass.ts pattern, this is architected as similar to the d_wood.cpp as possible. This makes it simpler to compare to the decompiled results.

A few functions are implemented, and the rendering is based heavily off of d_tree so it may not be completely correct. But basic model rendering and drop shadows are present.