Wind Waker / Twilight Princess: Fix sun PeekZ
We were testing if the sun's position was obscured by something closer. However, the game only places the sun 8000 units away, which is very easy to hit with one of the bigger maps. The original game seems to test if the depth buffer has *anything* in it at all. I can't recall exactly why I went with the projected Z thing, so just do this instead.