Fetch a Stage's Demo archive when it is loaded
The LBNK section of the Room info contains a u8 for each of the 12 layers. This int is the index of the Demo arc that the stage expects to be loaded for that layer (or 0xFF for no arc).

For instance, Outset Island (Room 44 of 'sea') specifies 0x02 in LBNK[0]. It expects Objects/Demo02.arc to be loaded. Demo events will then load demos using filenames that map into that arc.

In the game, this occurs just after dStage_dt_c_roomLoader() is called, during phase 2 of dScnRoom_Create().