Added correct room and layer information for all demos
They were extracted from the game using a modified version of the excellent wwrando/wwlib/stage_searcher.py script from LagoLunatic. Most of this data comes from the PLAY action of the PACKAGE actor in an event from a Stage's event_list.dat. This action has properties for the room and layer that these cutscenes are designed for. HOWEVER, this data is often missing.

It has been reconstructed by cross-referencing each Room's lbnk section (which points to a Demo*.arc file for each layer), the .stb files contained in each of those Objects/Demo*.arc files, and the FileName attribute from the event action.