Framework: Added fopAcM_searchFromName
This searches all processes based on the name. For the test, it uses base_process_class::mProfName, which is the index used to access the f_pc_profiles table. Currently in noclip we only store the pcName that comes out of the profile. So far, pcName and profName always seem to match, but the game frequently accesses both.

I'm now storing profName and added an assert if they ever differ. I think this will make it clearer in the future if there is another function that access profName.