Eyes now render in front of hair, but with correct scene depth
This is based off of the logic from Jasper's E4_WWZAT_b2 branch. Currently, we must manually set the translucency to false for the first eye alpha mask. When it's true (as it is in the BMD) noclip will always render it after the rest of the opaque body. I suspect something is not quite lining up with the game regarding how we're handling translucency sorting...