Wind Waker: Fix camera execution ordering problem
This was causing single frame glitches in demos when the camera was moved. Previously, the camera logic was executing before the demo manager. The demo would set the camera position, as well an actor position that was previously visible. The camera was expected to be moved that frame, allowing the actor to also move without issue.

But the actor logic was running after demo manager, so the actor would move off screen this frame while the camera would not. On the next frame the camera would pick up the position change.

The order needs to be such:
- Demo manager executes
- Camera executes
- Camera draws (final matrices are set)
- Everything else draws (using the new camera matrices)

This change makes dCamera_c a leafdraw process and adds it to the proper queues. One thing worth noting is that because of the way that the layer system has been simplified in Noclip, the camera must be created before d_s_play in order to be drawn first.