JSYSTEM: Introducing 'ResourceResolver' concept for resources which reference other resources
J2D BLO files reference other texture resources. The games need a way of resolving these references into the actual resources. A ResourceResolver is a simple callback of the form:

`ResourceResolver = (resType: string, resName: string) => any`

Note that the arc which contains those resources is not passed directly to the callback, because the BLO file does not know which arc it is being loaded from. Instead, the resolver is expected to have been created with the arc already specified. E.g. In Wind Waker this is handled by `dResControl_c.getResResolver(arcName: string): ResourceResolver` which returns a resolver function for the specified arc.

Passing the JKRArchive directly to the BLO was also considered. But in the case of Wind Waker, the BTI images that are referenced have already been loaded by dResCtrl_c, so having the BLO create them again would result in duplicate resources. The callback seems to be the best way to resolve this.