J2D: Introduce J2DAnchorPos to correct for wider aspect ratios
J2DScreens are typically 640x480. When rendering in a different aspect ratio, the height will scale to match but the new and old widths will differ. The new J2DAnchorPos determines at what X position the elements will be rendered. Either anchored in the center, or on the left/right side of the screen.

E.g. With J2DAnchorPos.Left an element that was 10% from the left edge of the original 640x480, it will be 10% from the left screen edge.