Wind Waker: Fix black bar issue when title demo ends
Previously, as soon as the Title demo ended the camera would reset to cinematic trimming mode (because `globals.scnPlay.demo.getName()` is null), then immediately transition back to default trimming.

Now the camera stores its trim height via `setTrimHeight()` which is reset on scene creation. This is closer to the game logic.