new deku (#747)
* Update to deku v0.18.0

* rust: Fix bit-endianness issue with new-deku

This was the cause of all of our WoW rendering failures with new-deku. Unfortunate we can't (yet?) use Deku for this part, but honestly I think the shift + mask is better than a loop anyway.

* rust: Map seek errors to deku errors

* rust: Fix Unity loading

The deku::byte_offset trick we do for alignment requires reader.bits_read to be correct, as that's the number we pad by. In our PackedVec implementations, we read a u8, and then later skip over it, which double-counts from the perspective of reader.bits_read. Fix this by adjusting after we read it.

* rust: Use std::io instead of deku::no_std_io

* rust: Update polymorph