Fix lighting on mDoExt_3DlineMat1_c
The main problem was that Noclip has added logic which disables diffuse lighting if the light attenuation is set to None. See generateLightDiffFn(). I don't believe this matches the original GX behavior.

The displaylists for 3Dline set diffuse to CLAMP and atten to NONE, which causes lighting to be enabled given the current behavior. This manually modifies the material to set atten to SPOT.

Additionally, we swizzle the lighting value from the r channel to rgb. Otherwise the lines appear to be lit by a red light. How does this work for other objects?