Wind Waker: Bridge (#755)
* Starting point for d_a_bridge

Ropes above and between planks are ignored. It's planks are forming the basic shape, but are linear and not yet following a curve.

* Fix up bridge curvature and plank scaling

* Fix sim/render plank positions

After simulation, all plank positions are offset (to a reducing degree) so that the first plank always lands on the start position

* The number of planks per bridge is now limited correctly

This sets the correct number of planks on the Outset bridge

* Bridge planks sway correctly

* Implement most of mDoExt_3Dline_c

Still some optimization work to do, and several errors. Lighting is currently fully red. The diffuse calculations don't seem to be working, and only the red channel of the light is being considered.

* The main suspension rope is rendered using the new mDoExt_3Dline_c

* Added fpcEx_Search(). Use to allow two half bridges to connect in the middle.

* Format d_a_bridge

* Format mDoExt_3DLine*

* Draw ropes down to planks

* Fix lighting on mDoExt_3DlineMat1_c

The main problem was that Noclip has added logic which disables diffuse lighting if the light attenuation is set to None. See generateLightDiffFn(). I don't believe this matches the original GX behavior.

The displaylists for 3Dline set diffuse to CLAMP and atten to NONE, which causes lighting to be enabled given the current behavior. This manually modifies the material to set atten to SPOT.

Additionally, we swizzle the lighting value from the r channel to rgb. Otherwise the lines appear to be lit by a red light. How does this work for other objects?

* Remove old TODO

* Parse 3DLine display lists into materials only once

* Set fog block

* Remove specular enable hack now that GX material handling has been fixed

See 55b3cb1781bc225012ff05be47fc78786786d971 for the fix

* Fix bridge crash in the case of multiple rooms being loaded at once

e.g. Dragon Roost Cavern

* Small cleanups

- Add missing 'public' specifier to mDoExt_3Dline_c.init()
- Don't use template render inst when only doing one draw call
- Don't use for..of in tight loops
- Use nicer TDDraw.beginDraw() interface

* Add access qualifiers to br_s

* Fix incorrect rotation applied to start and end rope positions

* Remove TODO. Confirmed third bit of BridgeFlags is NoRopes

Bridge is created without any support ropes (planks only)

* Reference 3DLine display lists via extra symbol data

* Parse display list into material once for each 3DLine

* Added `dKy_GxFog_tevstr_set` to fix 3DLine fog rendering

Fog values must come from the tev string, not the global env light

* Remove unnecessary double buffering from 3DLines, noclip handles this for us

* Chains on metal bridges now render

* Confirmed cLib_addCalc0 is implemented correctly

* Set the global env light to (1, 1, 1) instead of (1, 0, 0)

This fixes the lighting/TEV issue with mDoExt_3DlineMat1_c. Since it does not apply toon shading via ZAToon the diffuse light (previously red) is directly multiplied by the tex color. Without this fix all 3DLines would be tinted red.

* Small cleanups...

- Use colorNewFromRGBA8 when using 1 byte channel values
- Replace implementation of cLib_addCalc0 with simpler call to cLib_addCalc2
- Replace instances of ==/!= with ===/!==
- Small formatting fixes